feat: 迁移 ContentResponse 接口并更新相关类型定义

+++ packages/shared/src/utils/BookParser/parsers/TXTBookParser.ts
@@ -5,12 +5,13 @@ import type { BookChapter, IBookParser } from '../types';
 import { BaseBookParser } from './BaseBookParser';
 
 export default class TXTBookParser extends BaseBookParser implements IBookParser {
+  /** 章节正则 */
   private chapterPattern =
     /^第\s{0,4}[\d〇零一二两三四五六七八九十百千万壹贰叁肆伍陆柒捌玖拾佰仟]+?\s{0,4}(?:章|节(?!课)|卷|页|集|部|篇(?!张)).{0,40}$/;
 
   private async _getText(filePath: string): Promise<string> {
     const sourceFile = await fs.readFile(filePath);
-    const encoding = chardet.detect(sourceFile);
+    const encoding = chardet.detect(sourceFile as Uint8Array);
     const text = iconv.decode(sourceFile, encoding as string);
     return text;
   }
